refactor: consolidate audio/video components
