+++ src/components/Inputs/MediaRecord/index.js
@@ -1,0 +1,1 @@
+export { default } from './MediaRecord.vue';
