--- src/components/Inputs/WebAudioVideoRecord/index.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line
-export { default } from './AudioVideo.vue';
